scripts: add 5 new features
Add supporting for 4 new feature into scripts:

1). Possibility to use this sonic-scripts
   from same local machine where the sonic-buildimage is built/compiled
   but not only from cloud's github.
The feature is needed for development period
to easy compose this "sonic-scripts.git" with new and trial patches.

2). Separate Teralynx "ABT_*" directory name and Prestera "ABP_*"

3). Add submodules requrcive fetch all
When a submodule has merged PR (e.g. advanced) and has nes hash-number
but main sonic-build git is still not PR updated and pointing to previous hash,
the pointer cannot be updated by a patch since the main-git is not familiar
with this new hash.
Requrcive fetch-all resolves this constraint and the submodule hash update
by patch is working properly.
(requrcive fetch-all adds 12..15 seconds only to the build-time).

4). Disable script-debug/trace "set -e"

5). Build with 1 retry
Sometimes the "make..." is failed but re-run the same "make..."
command builds the SONiC projet successfuly (known SONiC build problem).
This feature does not abort the build on first failure but retries
(with forcing SONIC_BUILD_JOBS=1).

Signed-off-by: Yan Markman <ymarkman@marvell.com>