[utilities] restart mgmt over-usb interface if not RUNNING (#184)
PROBLEM:
Sometimes after "config reload" command the management
interface (eth0) connected over ETH-USB device is UP/RUNNING
on a remote peer side, but stays not-running on the local side.
This is well-known negotiation problem of ETH-over-USB devices.

SOLUTION:
If the management interface is a USB device and is not RUNNING
(/sys/class/net/eth0/operstate is not "up") at the end of the
"config reload" command
bring it down and up with delay to trigger renegotiation.

Signed-off-by: Yan Markman <ymarkman@marvell.com>