installer write to hung_task_timeout_secs always-ok
The original patch disables /proc/sys/kernel/hung_task_timeout_secs=0
for the ONIE-installer.
The hung_task_timeout presence depends upon Kernel-configuration
and not always present in Kernel. For example it is present on X86
but by-default does not present on ARM64.
It is desirable to have the patch-set/series for all ARCH-es.

=> make this writing always success even if file does not exist.

Signed-off-by: Yan Markman <ymarkman@marvell.com>